Move waitForCdTaskExit to cli pkg, and delay before retry
Also change provider delay before retry do exponential backoff
