+++ src/pkg/cli/tailAndMonitor.go
@@ -43,30 +43,17 @@ func TailAndMonitor(ctx context.Context, project *compose.Project, provider clie
 	go func() {
 		defer wg.Done()
 		// block on waiting for services to reach target state
-		for {
-			serviceStates, svcErr = WaitServiceState(svcStatusCtx, provider, targetServiceState, project.Name, tailOptions.Deployment, computeServices)
-			if svcErr != nil && isTransientError(svcErr) {
-				term.Debug("WaitServiceState failed with transient error, retrying:", svcErr)
-				continue
-			}
-			break
-		}
+		serviceStates, svcErr = WaitServiceState(svcStatusCtx, provider, targetServiceState, project.Name, tailOptions.Deployment, computeServices)
 		term.Debug("WaitServiceState stopped with", svcErr)
 	}()
 
 	go func() {
 		defer wg.Done()
 		// block on waiting for cdTask to complete
-		for {
-			err := client.WaitForCdTaskExit(ctx, provider)
-			if err != nil && isTransientError(err) {
-				term.Debug("WaitForCdTaskExit failed with transient error, retrying:", err)
-				continue
-			}
+		if err := WaitForCdTaskExit(ctx, provider); err != nil {
 			cdErr = err
 			// When CD fails, stop WaitServiceState
 			cancelSvcStatus(cdErr)
-			break
 		}
 		term.Debug("WaitForCdTaskExit stopped with", cdErr)
 	}()
